+++ site/content/release-notes.md
@@ -46,6 +46,16 @@ title = "Release Notes"
   `org.eclipse.hono.client.RegistrationClient` instances.
   The *setRegistrationServiceClient* and *getRegistrationServiceClient* methods have been
   renamed to *setRegistrationClientFactory* and *getRegistrationClientFactory* correspondingly.
+* The `org.eclipse.hono.service.AbstractProtocolAdapterBase` class now uses
+  `org.eclipse.hono.client.CredentialsClientFactory` instead of
+  `org.eclipse.hono.client.HonoClient` for creating
+  `org.eclipse.hono.client.CredentialsClient` instances.
+  The *setCredentialsServiceClient* and *getCredentialsServiceClient* methods have been
+  renamed to *setCredentialsClientFactory* and *getCredentialsClientFactory* correspondingly.
+* The `org.eclipse.hono.service.auth.device.UsernamePasswordAuthProvider` and the
+  `org.eclipse.hono.service.auth.device.X509AuthProvider` now accept a
+  `org.eclipse.hono.client.CredentialsClientFactory` instead of a
+  `org.eclipse.hono.client.HonoClient` in their constructors.
 
 ## 1.0-M1